Fix bug: trim according to least possible index for all channels
Originally, trimming was done based on the first channel, which means
that it only considered T-array 1.  However, if the impact was closer
to T-array 2 instead of 2, the first peak in channel 1 would be behind
the first peak in any channel in array 2.  Therefore, we need to
inspect every channel to find the minimum index and use that to trim
all channels.